+++ example/src/main.rs
@@ -1,6 +1,8 @@
 #![allow(dead_code)]
 mod ga;
 mod pso;
+mod ff;
+mod aco;
 
 fn main() {
 	ga::ga_example();
